Process multiple URLs on CLI
I've modified it to iterate over the argv array, as I usually throw multiple urls at youtube-dl.

It would be great if we could pass options to the underlying youtube-dl, as I usually use --xattr, etc. to keep the metadata as well. But that might get complicated, passing data around the muxing process... I'm not much of a php guy, so I can't be of much help.
Maybe, as an intermediate solution, the --write-info-json option could be hardcoded in, and then later, after muxing, the resulting sidecar file could be moved from /tmp to the final storage location, along with the output file...

But I'm only throwing feature requests at you here - sorry for that.
Very helpful script! Thank you for your work.